Fix stack setup and jump to _rust_start.
With this commit the stack pointer (%esp) is succesfully loaded with a hardcoded value (0x132000)
whereas before it was failing to find the _STACK_TOP symbol and load its value into %esp. However,
this commit does NOT fix the issue of pushing the address of the start info struct onto the stack
so that _rust_start receives it, as _rust_start is still seeing "start_info_ptr == 0x0".