Refactor early boot asm.
Moved stack and pgtable sections to .bss section for fun. The entry
function now sets up an initial identity map of the kernel with 2MiB
pages and attempts to enable paging, load the GDT, setup the stack and
finally jump to 64-bit mode. Currently experiencing a fault when trying
to write %cr0 to start paging.